Add not finished production
